wip combine login and splash
