+++ src/constants/router.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { createBrowserRouter } from 'react-router-dom';
 import { App } from '../components/app/app';
 import { Error } from '../components/error/error';
-import { Login } from '../components/login/login';
 import { isProduction } from '../utils/is-production/is-production';
 import { Splash } from '../components/splash/splash';
 import { Students } from '../components/students/students';
@@ -20,7 +19,6 @@ export const ROUTER = createBrowserRouter([
       { path: '/', element: <Splash /> },
       { path: 'admins', element: <Admins /> },
       { path: 'coaches', element: <Coaches /> },
-      { path: 'login', element: <Login /> },
       { path: 'students', element: <Students /> },
       { path: 'inactive-rejected', element: <InactiveRejected /> },
       { path: 'student-info', element: <StudentInfo /> },
